// web/src/lib/schemas.ts
export const IdSchema = z.uuidv7();
// ...
export const TransactionBase = z.object({
  merchant: z.string(),
  amount: z.number(),
  notes: z.string(),
  date: z.coerce.date(),
  userId: IdSchema,
  accountId: IdSchema,
  budgetItemId: IdSchema,
  type: z.enum(["credit", "debit"]),
  recurringTemplateId: z.nullish(IdSchema),
});

export const TransactionSchema = addIdAndDatetime(TransactionBase);
export type Transaction = z.infer<typeof TransactionSchema>;